remove: delete table root page
